# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/webhooks": {
      "post": {
        "description": "Creates a webhook for a workspace.",
        "operationId": "WebhookController_create_api/v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key used to safely retry the creation request. The same key cannot be reused with a different request payload.",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "example": "7b9f3e2a-5f1c-4f4b-9d8d-123456789abc"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook created successfully."
          },
          "400": {
            "description": "Validation failed."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "User is not allowed to manage webhooks."
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "Webhook creation conflicts with the current state."
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a webhook",
        "tags": [
          "Webhooks"
        ]
      },
      "get": {
        "description": "Retrieves all webhooks registered for the specified workspace.",
        "operationId": "WebhookController_findAll_api/v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhooks retrieved successfully."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": "Validation failed",
                  "instance": "/api/v1/tasks",
                  "errors": [
                    "email must be an email",
                    "password must be longer than or equal to 8 characters"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "User is not allowed to manage webhooks."
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The request conflicts with the current state of the resource",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get workspace webhooks",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Manager API",
    "description": "NestJS Task Manager API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank",
            "description": "URI reference identifying the problem type. Defaults to about:blank."
          },
          "title": {
            "type": "string",
            "example": "Bad Request",
            "description": "Short, human-readable summary of the problem."
          },
          "status": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code generated by the origin server."
          },
          "detail": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/tasks",
            "description": "URI reference identifying the specific occurrence."
          },
          "requestId": {
            "type": "string",
            "example": "01JXYZ123456789",
            "description": "Optional correlation/request identifier."
          },
          "errors": {
            "example": [
              "email must be an email",
              "password must be longer than or equal to 8 characters"
            ],
            "description": "Optional validation or additional error details.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
      "CreateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://example.com/webhooks/task-events",
            "description": "HTTPS endpoint that will receive webhook events."
          },
          "events": {
            "example": [
              "task.created",
              "task.updated",
              "task.deleted"
            ],
            "description": "Events that should trigger this webhook.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "url",
          "events"
        ]
      },
...
    }
  }
}